requirements.txt
# /wealth-advisor/requirements.txt

google-genai==1.29.0
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-run
google-auth
requests
gunicorn
pandas
flask
websockets
google-cloud-secret-manager
google-cloud-discoveryengine